requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic-settings==2.11.0
sqlalchemy==2.0.44
alembic==1.14.1
psycopg[binary]==3.3.3
pytest==8.4.2
httpx==0.28.1
pypdf==5.9.0
Pillow==11.3.0
pytesseract==0.3.13
python-multipart==0.0.20
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
